Separate legacy mongodb-based caching
